requirements.txt
tensorflow
opencv-python-headless
numpy
matplotlib
Pillow
uuid